requirements.txt
# AttentionX+ — Required Packages
# Install with: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.4

# ── ASR (Stage 1) ─────────────────────────────────────────────────────────────
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

# ── UserStyleVector / NLP (Stage 2 & 4) ──────────────────────────────────────
sentence-transformers==2.7.0
transformers==4.40.0
numpy>=1.24.0
scipy>=1.11.0

# ── Sentiment / Highlight Detection (Stage 3) ─────────────────────────────────
librosa==0.10.1
soundfile==0.12.1
vaderSentiment==3.3.2
textblob==0.18.0

# ── Video Processing (Stage 5) ────────────────────────────────────────────────
moviepy==1.0.3
imageio==2.34.0
imageio-ffmpeg==0.4.9
Pillow>=10.0.0

# ── Face Detection for Smart Crop (Stage 5) ───────────────────────────────────
mediapipe==0.10.11
opencv-python-headless==4.9.0.80

# ── Anthropic API (caption rewriting fallback) ────────────────────────────────
anthropic>=0.25.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
